[FedCM] Fix some chromedriver tests without 3PCs
When third party cookies are disabled, FedCM does not work unless the
IDP Sign-In Status API is enabled. In order to reduce test failures
seen when disabling third party cookies, we enable the API on a these
tests. We also need to mark signin to ensure the FedCM dialog is shown.

Bug: 1484833

Change-Id: Ia8de19f9558717f7eb3fe676da0c1168045fbb03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4889802
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Nicolás Peña <npm@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1202055}