Fix ChromeDriver license generation warnings
The license generation step of ChromeDriver target issued messages that
the third party dependencies of selenium-atoms were not listed in
SPECIAL_CASES of licenses.py. This commit removes the
--extra-third-party-dirs argument from the licenses.py invocation.
As a result the script generates the license only basing on the
information in //third_party/selenium-atoms/README.chromium which
includes all the necessary third party notices.

This commit also opts to use MIT license for Sizzle that allows multiple
dependencies.

Bug: chromedriver:4735
Change-Id: Id3b0b8eab6b04f63516014d2d564f0bfaff7f22c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5339957
Commit-Queue: Vladimir Nechaev <nechaev@chromium.org>
Reviewed-by: Maksim Sadym <sadym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1268337}