Minimize wait time in ChromeImpl::SetWindowBounds
With this commit ChromeImpl::SetWindowBounds waits with exponential
backoff until the window animation is finished only on condition that
the window bounds have not yet changed to the desired ones. Otherwise
the bounds deemed to be successfully applied.
This change must reduce the current latency in WPT setup.

Bug: chromedriver:4642
Change-Id: I508a9b3e21c23dcc4948ce25d67568ef00d03883
Validate-Test-Flakiness: skip
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5366457
Reviewed-by: Maksim Sadym <sadym@chromium.org>
Commit-Queue: Vladimir Nechaev <nechaev@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1272781}