Wait for navigation while finding elements
With this commit ChromeDriver waits for a navigation detected by the
remote end while searching for element or elements. Before this commit
ChromeDriver used to return a non-standard error to the user in Release
builds or crash with DCHECK in Debug builds.
The fix is tested with unit tests only as inegration tests are difficult
to design to reproduce the problem more or less reliably.

Bug: chromedriver:4771
Change-Id: I46b2d392b12c60aebc24386bc1a922f3f8985cdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5562089
Auto-Submit: Vladimir Nechaev <nechaev@chromium.org>
Commit-Queue: Maksim Sadym <sadym@chromium.org>
Reviewed-by: Maksim Sadym <sadym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1305054}