Remove uses of variable-length arrays
Variable-length arrays are not part of the C++ standard, and are
dangerous to use because they allocate a potentially-unbounded
amount of data on the stack. In addition, they trigger a clang warning
(-Wvla-extension) which is currently suppressed in Chromium builds.

This simply removes a warning by marking the length of the array as
`constexpr`, allowing the compiler to figure out that its length is
static.

Bug: 349656479
Change-Id: I19434b78b1d15578826cf426557ff1c1e93e7352
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5738737
Reviewed-by: Vladimir Nechaev <nechaev@chromium.org>
Commit-Queue: Devon Loehr <dloehr@google.com>
Reviewed-by: Maksim Sadym <sadym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1332914}