Unfork the WebSocketEncoder implementations
This change addresses the stalled migration of the WebSocketEncoder
from `//net/server` to `//services/network/public/cpp/server`.

The initial plan was to migrate to
`//services/network/public/cpp/server` to consolidate WebSocket
handling, but progress on this migration stalled, resulting in
duplicated and fragmented implementations.

Additionally, recent updates exclusively applied to `//net/server`
have exacerbated the need to revert and centralize the logic back to
this location.

- Removes WebSocketEncoder code from
  `//services/network/public/cpp/server`.

- Redirects all references back to the consolidated implementation in
  `//net/server`.

- Extracts WebSocket parse result enums into `web_socket_parse_result.h`
  to promote broader accessibility and reuse.

- Encountered issues with external references using
  `//net/server/BUILD.gn`, leading to the decision to centralize all
  files under `//net/server` within the `//net/BUILD.gn` to simplify
  and streamline build configurations.

By reverting to `//net/server`, this change eliminates the fragmentation
and inconsistency in WebSocket handling that had developed due to the
incomplete migration and unilateral updates.

Centralizing WebSocketEncoder in `//net/server` ensures that all future
updates are uniformly applied, enhancing the maintainability and
stability of the WebSocket functionality within the Chromium project.

Bug: 40205564
Change-Id: Ic890ecd619f7df6ab17a6b58d9d23e291644ac9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5768214
Reviewed-by: Maksim Sadym <sadym@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Reviewed-by: Huanpo Lin <robertlin@chromium.org>
Commit-Queue: Ryosuke Miyake <rysk@google.com>
Cr-Commit-Position: refs/heads/main@{#1343880}