Fix mouse clicks in ChromeDriver
The implementation before this commit always sent mouse events to the
top most frame in hope that it would be routed to the appropriate frame.
This turned out to be racy in the case with remote frames.
The current commit changes this behavior so that mouse events are sent
to the highest same source ancestor of the target frame. This ensures
that mouse events are properly routed even if the frame tree is not yet
assembled.

Bug: 42323535
Change-Id: Ia6695bdfcb84cd185e29b65875f9f020293f09aa
Validate-Test-Flakiness: skip
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5836968
Reviewed-by: Maksim Sadym <sadym@chromium.org>
Commit-Queue: Vladimir Nechaev <nechaev@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1356417}