Fix testScriptNavigateRemoteToLocal
The browser side does not guarantee that some scripts will return the
result. This concerns the scripts making a navigation or closing the
window. ChromeDriver ensures that in such cases the standard compliant
ScriptTimeout error is returned to the user. This signals the user the
ambiguous status of the command and hints that the test needs to be
refactored to avoid such an ambiguity.

Bug: 372883701
Change-Id: Iaef0e13e6ce38ac5017ab984f4a2b486e29885a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5934265
Commit-Queue: Vladimir Nechaev <nechaev@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1369269}