ChromeDriver: WebDriver BiDi sub-client should inherit ChromeDriver permissions
This change is needed for WebDriver BiDi clients to test extensions that
require permissions that the WebDriver client usually has. This change
allows the connection inside the WebDriver BiDi WebContents to inherit
the WebDriver client permissions. The permissions are only added
if the client specified `--enable-unsafe-extension-debugging` in
Chrome arguments.

Bug: 393500934
Change-Id: Ia8e2501095b9afb5d7c783031a376eafe244e517
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6288627
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Maksim Sadym <sadym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1424990}