[chromedriver] Allow clients to stop browser process gracefully
Add a Chrome-specific capability `quitGracefully` that forces
chromedriver to attempt to stop the browser with `Browser.close` or
SIGTERM. wptrunner will enable this capability when used with the `rr`
debugger (`rr record chrome ...`), which needs time to dump the
recording. The runner doesn't want to manage `--user-data-dir` or
`--log-net-log` itself, so the existing code path for graceful quit
is insufficient.

SIGKILL remains the default behavior so that quitting remains fast for
clients that don't opt-in.

Bug: 339644807
Change-Id: I8b4286bb188f36d0c786e78bf6257c516c91f865
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6269180
Commit-Queue: Jonathan Lee <jonathanjlee@google.com>
Reviewed-by: Maksim Sadym <sadym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1427507}