Fix ClearSetDisplayFeatures not working correctly on cft bots
We can simplify the test, after all the functionality of the Viewport Segments API is tested elsewhere. The test is to ensure that ClearDisplayFeatures resets the viewport segments to the default values (which is the viewport size). So, let's just directly test that after calling ClearDisplayFeatures, the viewport segments are the same as the viewport size.

Bug: 417195296
Change-Id: I6e0560680f99347cd997242b68b5f0656ca1fa7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6535067
Reviewed-by: Alex N. Jose <alexnj@chromium.org>
Commit-Queue: Alexis Menard <alexis.menard@intel.com>
Cr-Commit-Position: refs/heads/main@{#1458910}