Fix chromedriver crash when dismiss the glic fre dialog
This CL fixed a regression for https://crbug.com/410599467 that when
click on the glic fre dismiss button, the chromedriver might crash.

The reason was that, when `ExecuteClickElement` the FRE continue button,
the FRE webview was closed, and the click action could return with
kTargetDetached status. If the status been overriden as `kOk` and
returned to `ExecuteWindowCommand`[1], web_view will be invalid and
web_view->WaitForPendingNavigations will fail with "Access Violation".

This CL also changed the switch process for --glic-automation and
--glic-always-open-fre that we can test the FRE instead of bypass it.

The test is flaky without the fix, that please run the test multiple
times:

```
vpython3 chrome\test\chromedriver\test\run_py_tests.py \
--chromedriver=<LOCAL_BUILD> \
--filter=__main__.ChromeDriverTest.testWebviewDetactedDuringClick \
--repeat=20
```

or check the test results from CQ (without the fix):
https://ci.chromium.org/ui/p/chromium/builders/try/linux-rel/2220939/test-results?sortby=&groupby=&q=ChromeDriverTest.testWebviewDetactedDuringClick+
https://ci.chromium.org/ui/p/chromium/builders/try/mac-rel/2187476/test-results?sortby=&groupby=&q=ChromeDriverTest.testWebviewDetactedDuringClick
https://ci.chromium.org/ui/p/chromium/builders/try/win-rel/1019128/test-results?sortby=&groupby=&q=ChromeDriverTest.testWebviewDetactedDuringClick

[1]
https://source.chromium.org/chromium/chromium/src/+/main:chrome/test/chromedriver/window_commands.cc;l=763;drc=f5aefb2a059239ebf2475d6ebfebee9af974eced

Bug: 410599467
Change-Id: If245aea5e068edcbcfac6c78487063d277929a3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6538867
Commit-Queue: Kuan Huang <kuanhuang@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: Maksim Sadym <sadym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1460241}