chromedriver: Don't use ancestor overflow for in-view center point
As per the webdriver spec [1] the in-view center point should only be
the intersection of element's bounds and the viewport. Any sort of
clipping that may occur due to an ancestor specifying overflow:hidden
should not affect the in-view center point calculation.

The implementation of PerformActions pointerMove with an element origin
was returning an unexpected center point value due to this extra
overflow check. So remove that so that it is consistent with the spec
and other cases like in element.Click() and MouseMoveTo().

[1] https://w3c.github.io/webdriver/#dfn-center-point

Bug: 42322257
Change-Id: Ib3cef0f564c3714713113c4c8f9f3453ad3406e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6735230
Reviewed-by: Alex N. Jose <alexnj@chromium.org>
Commit-Queue: Orko Garai <orko@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1490887}