Avoids stack trace print for kTabCrashed status
Printing a ChromeDriver stack trace when a tab is crashed produces user
confusion and is not very useful for the specific failure.

ChromeDriver was initially not capable of detecting tab crashes and was
collecting a stack trace for every non-OK status code. However, at a
later point, kTabCrashed status code was implemented but the stack trace
collection was not removed for it. This CL avoids collecting
ChromeDriver and printing it when kTabCrashed status is detected. Also
includes a regression test that confirms that a renderer crash stack
trace is present in the log, and no ChromeDriver stack trace is
collected or printed on stdout.

Bug: 395131239
Change-Id: I25095317b38cb1a58e37f1b8df86b4ec09076ece
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6841669
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1500624}